requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
cryptography==41.0.7  # Required for secure MySQL connections